+++ pypeit/calibrations.py
@@ -67,6 +67,7 @@ class Calibrations(object):
         msbias (:class:`pypeit.images.buildimage.BiasImage`):
         msdark (:class:`pypeit.images.buildimage.DarkImage`):
         msbpm (`numpy.ndarray`_):
+        alignments (:class:`pypeit.alignframe.Alignments`):
         wv_calib (:obj:`dict`):
         slits (:class:`pypeit.slittrace.SlitTraceSet`):
 
@@ -150,8 +151,7 @@ class Calibrations(object):
 
         self.msarc = None
         self.mstilt = None
-        self.msalign = None
-        self.alignment = None
+        self.alignments = None
         self.msbias = None
         self.msdark = None
         self.msbpm = None
@@ -301,93 +301,47 @@ class Calibrations(object):
         """
         Load or generate the alignment frame
 
-        Requirements:
-           master_key, det, par
+        Requires: :attr:`slits`, :attr:`det`, :attr:`par`
 
         Returns:
-            ndarray or str: :attr:`align`
+            :class:`pypeit.alignframe.Alignments`:
 
         """
         # Check for existing data
-        if not self._chk_objs(['msbpm', 'tslits_dict']):
-            msgs.error("Don't have all the objects")
+        if not self._chk_objs(['msbpm', 'slits']):
+            msgs.error('Must have the bpm and slits to make the alignments!')
 
         # Check internals
         self._chk_set(['det', 'calib_ID', 'par'])
 
         # Prep
-        align_files = self._prep_calibrations('align')
-        #align_rows = self.fitstbl.find_frames('align', calib_ID=self.calib_ID, index=True)
-        #self.align_files = self.fitstbl.frame_paths(align_rows)
-        #self.master_key_dict['align'] \
-        #        = self.fitstbl.master_key(align_rows[0] if len(align_rows) > 0 else self.frame,
-        #                                  det=self.det)
-        masterframe_name = masterframe.construct_file_name(
-            buildimage.AlignImage, self.master_key_dict['align'], master_dir=self.master_dir)
+        align_files, self.master_key_dict['align'] = self._prep_calibrations('align')
 
-        # Previously cahded?
-        if self._cached('align', self.master_key_dict['align']):
-            # Previously calculated
-            self.msalign = self.calib_dict[self.master_key_dict['align']]['align']
-        elif os.path.isfile(masterframe_name) and self.reuse_masters:
-            self.msalign = buildimage.AlignImage.from_file(masterframe_name)
-        else:
-            # Instantiate with everything needed to generate the image (in case we do)
-            #self.alignFrame = alignframe.AlignFrame(self.spectrograph, files=self.align_files,
-            #                                  det=self.det, msbias=self.msbias,
-            #                                  par=self.par['alignframe'],
-            #                                  master_key=self.master_key_dict['align'],
-            #                                  master_dir=self.master_dir,
-            #                                  reuse_masters=self.reuse_masters)
-            self.align = buildimage.buildimage_fromlist(self.spectrograph, self.det,
-                                                         self.par['alignframe'],
-                                                         align_files, bias=self.msbias, bpm=self.msbpm)
+        masterframe_filename = masterframe.construct_file_name(alignframe.Alignments,
+                                                               self.master_key_dict['align'],
+                                                               master_dir=self.master_dir)
 
-            # Load the MasterFrame (if it exists and is desired)?
-            #self.msalign = self.alignFrame.load()
-            #if self.msalign is None:  # Otherwise build it
-            #    msgs.info("Preparing a master {0:s} frame".format(self.alignFrame.master_type))
-            #    self.msalign = self.alignFrame.build_image(bias=self.msbias, bpm=self.msbpm)
-            #    # Need to set head0 here, since a master align frame loaded from file will have head0 set.
-            #    self.msalign.head0 = self.alignFrame.build_master_header(steps=self.alignFrame.process_steps,
-            #                                                         raw_files=self.alignFrame.file_list)
-            #   # Save to Masters
-            # Save to Masters
-            self.msalign.to_master_file(self.master_dir, self.master_key_dict['align'],  # Naming
-                                       self.spectrograph.spectrograph,  # Header
-                                       steps=self.msalign.process_steps,
-                                       raw_files=align_files)
+        # Reuse master frame?
+        if os.path.isfile(masterframe_filename) and self.reuse_masters:
+            self.alignments = alignframe.Alignments.from_file(masterframe_filename)
+            self.alignments.is_synced(self.slits)
+            return self.alignments
 
-            # Store the alignment frame
-            self._update_cache('align', 'align', self.msalign)
+        msalign = buildimage.buildimage_fromlist(self.spectrograph, self.det, self.par['alignframe'], align_files,
+                                                 bias=self.msbias, bpm=self.msbpm)
 
-        # Check if the alignment dictionary exists
-        if self._cached('align_dict', self.master_key_dict['align']) \
-                and self._cached('wtmask', self.master_key_dict['align']):
-            self.align_dict = self.calib_dict[self.master_key_dict['align']]['align_dict']
-        else:
-            # Extract some header info needed by the algorithm
-            binning = self.spectrograph.get_meta_value(self.align_files[0], 'binning')
+        # Extract some header info needed by the algorithm
+        binning = self.spectrograph.get_meta_value(align_files[0], 'binning')
 
-            # Instantiate
-            self.alignment = alignframe.Alignment(self.msalign, self.tslits_dict, self.spectrograph,
-                                                  self.par['alignment'],
-                                                  det=self.det, binning=binning,
-                                                  master_key=self.master_key_dict['align'],
-                                                  master_dir=self.master_dir,
-                                                  reuse_masters=self.reuse_masters,
-                                                  qa_path=self.qa_path, msbpm=self.msbpm)
+        # Instantiate
+        alignment = alignframe.TraceAlignment(msalign, self.slits, self.spectrograph, self.par['alignment'],
+                                              det=self.det, binning=binning, qa_path=self.qa_path, msbpm=self.msbpm)
+        # Run
+        self.alignments = alignment.run(show=self.show)
+        # Save to Masters
+        self.alignments.to_master_file(masterframe_filename)
 
-            # Master
-            self.align_dict = self.alignment.load()
-            if self.align_dict is None:
-                self.align_dict = self.alignment.run(self.show)
-                self.alignment.save()
-
-            # Save & return
-            self._update_cache('align', 'align_dict', self.align_dict)
-
-        return self.msalign, self.align_dict
+        return self.alignments
 
     def get_bias(self):
         """
@@ -868,7 +822,7 @@ class IFUCalibrations(Calibrations):
 
         """
         # Order matters!
-        return ['bias', 'bpm', 'arc', 'tiltimg', 'slits', 'wv_calib', 'tilts', 'align', 'flats', 'wave']
+        return ['bias', 'dark', 'bpm', 'arc', 'tiltimg', 'slits', 'wv_calib', 'tilts', 'align', 'flats']
 
 
 def check_for_calibs(par, fitstbl, raise_error=True, cut_cfg=None):