+++ pypeit/spectrographs/keck_nires.py
@@ -82,6 +82,7 @@ class KeckNIRESSpectrograph(spectrograph.Spectrograph):
         par['calibrations']['slitedges']['trace_thresh'] = 10.
         par['calibrations']['slitedges']['fit_min_spec_length'] = 0.4
         par['calibrations']['slitedges']['left_right_pca'] = True
+        par['calibrations']['slitedges']['fwhm_gaussian'] = 4.0
 
         # Tilt parameters
         par['calibrations']['tilts']['tracethresh'] =  10.0
@@ -177,15 +178,15 @@ class KeckNIRESSpectrograph(spectrograph.Spectrograph):
             # No pinhole or bias frames
             return np.zeros(len(fitstbl), dtype=bool)
         if ftype == 'standard':
-            return good_exp & (fitstbl['idname'] == 'Object')
+            return good_exp & ((fitstbl['idname'] == 'object') | (fitstbl['idname'] == 'Object'))
         if ftype == 'dark':
             return good_exp & (fitstbl['idname'] == 'dark')
         if ftype in ['pixelflat', 'trace']:
             return fitstbl['idname'] == 'domeflat'
         if ftype in 'science':
-            return good_exp & (fitstbl['idname'] == 'Object')
+            return good_exp & ((fitstbl['idname'] == 'object') | (fitstbl['idname'] == 'Object'))
         if ftype in ['arc', 'tilt']:
-            return good_exp & (fitstbl['idname'] == 'Object')
+            return good_exp & ((fitstbl['idname'] == 'object') | (fitstbl['idname'] == 'Object'))
         return np.zeros(len(fitstbl), dtype=bool)