+++ pypeit/tests/test_scripts.py
@@ -3,6 +3,7 @@ Module to run tests on scripts
 """
 import os
 import shutil
+import numpy as np
 
 import matplotlib
 matplotlib.use('agg')  # For Travis
@@ -10,7 +11,7 @@ matplotlib.use('agg')  # For Travis
 from astropy.io import fits
 
 from pypeit.scripts import setup, show_1dspec, coadd_1dspec, chk_edges, view_fits, chk_flats
-from pypeit.scripts import trace_edges, run_pypeit, ql_mos, show_2dspec
+from pypeit.scripts import trace_edges, run_pypeit, ql_mos, show_2dspec, tellfit, flux_setup
 from pypeit.tests.tstutils import dev_suite_required, cooked_required
 from pypeit import edgetrace
 from pypeit import ginga
@@ -185,5 +186,6 @@ def test_coadd1d_2():
     # Clean up
     os.remove(parfile)
     os.remove(coadd_ofile)
+
 
 # TODO: Include tests for coadd2d, sensfunc, flux_calib